♻️ Move setup, builder and verifier methods to separate classes for test utils
